new obs: function prototype

+++ include/libswiftnav/track.h
@@ -243,5 +243,5 @@ int nav_meas_cmp(const void *a, const void *b);
 u8 tdcp_doppler(u8 n_new, navigation_measurement_t *m_new,
                 u8 n_old, navigation_measurement_t *m_old,
                 navigation_measurement_t *m_corrected, double dt);
-bool calculate_loss_of_lock(u32 dt, u32 prev_lock_time, u32 curr_lock_time);
+bool calculate_loss_of_lock(double dt, u32 prev_lock_time, u32 curr_lock_time);
 #endif /* LIBSWIFTNAV_TRACK_H */
